move XYDataCollector from dvrp to ev
